refactor (vector): rename GeoJSON2Feature to GeoJSON2Features
